выносит CHARACTERS в отдельный файл
